+++ packages/vt/src/layer/vector/Vector3DLayerRenderer.js
@@ -1303,6 +1303,11 @@ class Vector3DLayerRenderer extends CanvasCompatible(LayerAbstractRenderer) {
         }
         this.markRebuild();
         redraw(this);
+        this.layer.fire('removegeo', {
+            'type': 'removegeo',
+            'target': this,
+            'geometries': geometries
+        });
     }
 
     onGeometrySymbolChange(e) {
@@ -1477,7 +1482,7 @@ class Vector3DLayerRenderer extends CanvasCompatible(LayerAbstractRenderer) {
             this._lineMeshes && this._lineMeshes.length && this._linePainter.needRebuildOnGometryPropertiesChanged() ||
             this.meshes && this.meshes.length && this.painter.needRebuildOnGometryPropertiesChanged()) {
             this.markRebuild();
-        } else {
+        } else if (this.meshes) {
             this.painter.onFeatureChange(this.features[uid], this.meshes);
         }