+++ packages/maptalks-gpu/CHANGELOG.md
@@ -1,5 +1,36 @@
 # maptalks-gl
 
+## 0.118.1
+
+### Patch Changes
+
+- fix DrawToolLayer
+- Updated dependencies
+  - @maptalks/gpu@0.118.1
+  - @maptalks/3dtiles@0.118.1
+  - @maptalks/gltf-layer@0.118.1
+  - @maptalks/video-layer@0.118.1
+  - maptalks@1.7.1
+  - @maptalks/transform-control@0.118.1
+  - @maptalks/vt@0.118.1
+
+## 0.118.0
+
+### Minor Changes
+
+- fixes and features
+
+### Patch Changes
+
+- Updated dependencies
+  - @maptalks/gpu@0.118.0
+  - @maptalks/3dtiles@0.118.0
+  - @maptalks/gltf-layer@0.118.0
+  - @maptalks/video-layer@0.118.0
+  - maptalks@1.7.0
+  - @maptalks/transform-control@0.118.0
+  - @maptalks/vt@0.118.0
+
 ## 0.117.1
 
 ### Patch Changes