+++ packages/traffic/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @maptalks/traffic
 
+## 0.118.1
+
+### Patch Changes
+
+- fix DrawToolLayer
+- Updated dependencies
+  - @maptalks/gltf-layer@0.118.1
+
+## 0.118.0
+
+### Minor Changes
+
+- fixes and features
+
+### Patch Changes
+
+- Updated dependencies
+  - @maptalks/gltf-layer@0.118.0
+
 ## 0.117.1
 
 ### Patch Changes